services/resource/consumer/requirements.txt
boto3==1.37.20
botocore==1.37.20
bpy==4.2.0
cryptography==44.0.2
dotenv==0.9.9
Flask_Cors==5.0.1
Flask==3.1.0
google.genai==1.19.0
openai==1.77.0
pandas==2.2.2
pillow==11.1.0
pillow-heif==0.20.0
prometheus_client==0.21.1
pypdf2==3.0.1
python-json-logger==3.3.0
python-magic==0.4.27
rembg==2.0.59
SQLAlchemy==2.0.25
trimesh==4.6.5
Twisted==24.11.0